+++ server-go/assets/embed.go
@@ -7,3 +7,6 @@ var DemoAssets embed.FS
 
 //go:embed firmware
 var FirmwareAssets embed.FS
+
+//go:embed default-question-image.svg
+var DefaultQuestionImage []byte